+++ src/Anroots/Pgca/Cli/Command/Rules/ListCommand.php
@@ -3,7 +3,9 @@
 namespace Anroots\Pgca\Cli\Command\Rules;
 
 use Anroots\Pgca\Cli\ContainerAwareCommand;
+use Anroots\Pgca\Rule\RuleInterface;
 use Symfony\Component\Console\Helper\Table;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
@@ -14,6 +16,7 @@ class ListCommand extends ContainerAwareCommand
     protected function configure()
     {
         $this->setName('rules:list')
+            ->addArgument('category', InputArgument::OPTIONAL, 'Only list rules belonging to the specified category')
             ->setDescription('List all available rules');
     }
 
@@ -28,9 +31,27 @@ class ListCommand extends ContainerAwareCommand
         $table->setHeaders(['Name', 'Category']);
 
         foreach ($this->rules as $rule) {
+            if ($this->isRuleSkipped($input, $rule)) {
+                continue;
+            }
+
             $table->addRow([$rule->getName(), $rule->getCategory()]);
         }
 
         $table->render();
+    }
+
+    /**
+     * @param InputInterface $input
+     * @param RuleInterface $rule
+     * @return bool
+     */
+    private function isRuleSkipped(InputInterface $input, RuleInterface $rule)
+    {
+        if ($input->getArgument('category') === null) {
+            return false;
+        }
+
+        return strtolower($rule->getCategory()) !== strtolower($input->getArgument('category'));
     }
 }